Create symbolic link to /tshock/ServerLog.txt
TShock writes a log file directly into the /tshock directory, and the
path is not configurable. Added a symbolic link into /world, so the file
will be created in the mounted volume. This enables the use of the -u
parameter in Docker to run the server as an unprivileged user.