Regress to lastest compatible version of mono
I built a docker image with every single version of full/slim mono since the last known working version and found that the Vanilla docker image fails starting 6.12.0.117, regardless of slim or not. As such I recommend TSHOCK use the current 6.12.0.122-slim since it's known to work fine. and VANILLA should use 6.10.0.104-slim as that is the last good version of mono for that build. @ryansheehan Thank you for your great work on the docker image, sorry for the typo in my last submission. I didn't catch the change upstream in the url and didn't have a pipeline setup to confirm build.