Remove '-slim' modifier from TSHOCK's mono version
This patch should resolve https://github.com/ryansheehan/terraria/issues/73

per #73 & fpiesche: Tshock requires 'fsharp' which is not contained in -slim versions of mono. 
going forward slim versions of mono should not be used, so as to maintain mysql compatibility. 

If the ~100mb image savings are needed by a select few, suggest opening a ticket for interest and perhaps creating a tshock-slim version that forgoes complete compatibility for size; however a change to add fsharp to fix the known issue with slim should be introduced before the new image is released.